fix(cvector.h): delete resize method
Because I don't have any idea to add an unknown type value into array.
If it initialzed as NULL, I must edit the pointer in underlying array outside in future,
that means the return type of `at` method will be pointer to pointer to void,
and it will make accessing a element value difficult.